backend_cloudfunctions/gemini-hackathon-insta-influencer /requirements.txt
llama-index
google-generativeai>=0.3.0
matplotlib
qdrant_client
trulens-eval
google-cloud-storage